Completely changed the way the simulator works
The listener is completely split up in several functions. The buffering is more on the background and every frame there is a possibility to interupt the simulation. The data is also transformed such that it is nice for the front end to handle. Next to that a lot of documentation is added.